+++ changelog.md
@@ -21,6 +21,7 @@
 - *added* Kompatibilität zu HE Quicksave Unterdrückung
 - *added* Kompatibilität zu SetResourceAmount
 - *added* Kompatibilität zu Funktionen des Interfaces
+- *added* Kompatibilität zu API.GetControllingPlayer
 - *added* API.UseSingleStop(_Flag) zu BuildingUI Modul hinzugefügt
 - *added* API.UseSingleReserve(_Flag) zu BuildingUI Modul hinzugefügt
 - *added* API.UseDowngrade(_Flag) zu BuildingUI Modul hinzugefügt